feat(switchdash): flat agent list in the sidebar (CHOO-1440)
The sidebar renders a flat list of agents — no location grouping, no
representative agent, no nested subagents. Each agent is a first-class row
(SidebarAgentItem) that opens the agent's page, starts sessions launched as its
own definition, and nests its sessions (grouped by room) underneath. Both the
agent-focused and room-focused views iterate agents.

- Delete location-item.tsx, sidebar-subagent-list.tsx, sidebar-dnd.tsx and the
  isSubagentSession helper.
- Per-agent delete (useConfirmDeleteAgent takes an explicit agentId).

Whole workspace green: typecheck, lint, tests (shared 7, core 156, plugins 38,
app 1279).
