+++ dash/apps/switchdash-desktop/src/renderer/features/sidebar/sidebar-grouped-list.tsx
@@ -1,36 +1,58 @@
-import { DndContext } from '@dnd-kit/core';
-import { SortableContext, verticalListSortingStrategy } from '@dnd-kit/sortable';
 import { observer } from 'mobx-react-lite';
 import { Fragment, useEffect } from 'react';
 import { agentsStore } from '@renderer/features/locations/stores/agents-store';
+import type { LocationStore } from '@renderer/features/locations/stores/location';
 import type { SessionStore } from '@renderer/features/sessions/stores/session-store';
 import { switchRoomsStore as roomConnectionsStore } from '@renderer/features/switch-rooms/switch-rooms-store';
 import { switchRoomsStore } from '@renderer/features/switch-servers/switch-rooms-store';
 import { switchServersStore } from '@renderer/features/switch-servers/switch-servers-store';
 import { sidebarStore } from '@renderer/lib/stores/app-state';
-import { SidebarLocationItem } from './location-item';
+import type { Agent } from '@shared/core/agents/agents';
+import { SidebarAgentItem, agentExpandKey } from './agent-item';
 import { SidebarSessionItem } from './session-item';
-import { makeDndId, SortableBranch, SortableLeaf, useSidebarDnd } from './sidebar-dnd';
 import {
   groupByRoom,
-  isSubagentSession,
   openRoomInGateway,
   openRoomInMessagingApp,
   RoomRow,
   roomLabel,
 } from './sidebar-room-grouping';
-import { agentRoomGroupKey, roomViewGroupKey } from './sidebar-store';
-import { SidebarSubagentList } from './sidebar-subagent-list';
+import { agentRoomGroupKey, roomViewGroupKey, UNASSIGNED_ROOM_KEY } from './sidebar-store';
 
-/** dnd container ids. Each identifies a reorderable sibling set. */
-const AGENTS_CONTAINER = 'agents';
-const ROOMS_CONTAINER = 'rooms';
-/** Sessions of one agent's room group (agent-focused view). */
-const agentSessionsContainer = (locationId: string, roomKey: string): string =>
-  `as:${locationId}|${roomKey}`;
-/** Sessions of one agent within a room (room-focused view). */
-const roomSessionsContainer = (roomKey: string, locationId: string): string =>
-  `rs:${roomKey}|${locationId}`;
+/** The provider definition a session was launched as, if any. */
+function sessionDefinitionName(session: SessionStore): string | undefined {
+  return 'subagentName' in session.data ? session.data.subagentName : undefined;
+}
+
+/** An agent paired with its (mounted) location, for the flat sidebar list. */
+type AgentEntry = { agent: Agent; location: LocationStore };
+
+/**
+ * The flat list of agents in the active-server scope, newest first. switchdash
+ * shows agents as a flat list — not grouped by directory (CHOO-1440).
+ */
+function scopedAgents(): AgentEntry[] {
+  const entries: AgentEntry[] = [];
+  for (const location of sidebarStore.filteredLocations) {
+    for (const agent of agentsStore.byLocation.get(location.id) ?? []) {
+      entries.push({ agent, location });
+    }
+  }
+  return entries.sort(
+    (a, b) =>
+      b.agent.createdAt.localeCompare(a.agent.createdAt) || a.agent.name.localeCompare(b.agent.name)
+  );
+}
+
+/** An agent's visible sessions: its location's sessions launched as its definition. */
+function agentSessions(entry: AgentEntry): SessionStore[] {
+  const all = sidebarStore.visibleSessionsForLocation(entry.location.id);
+  const def = entry.agent.definitionName;
+  return all.filter((session) => {
+    const name = sessionDefinitionName(session);
+    return def == null ? !name : name === def;
+  });
+}
 
 export const SidebarGroupedList = observer(function SidebarGroupedList() {
   // Live session→room connections + room names live on the server; pull them
@@ -63,221 +85,154 @@ export const SidebarGroupedList = observer(function SidebarGroupedList() {
   );
 });
 
-const AgentFocusedTree = observer(function AgentFocusedTree() {
-  const locations = sidebarStore.filteredLocations;
-  // Populated during render (each rendered group registers its ordered ids) so
-  // the drag handler can resolve a container's siblings on drop.
-  const containers: Record<string, string[]> = {
-    [AGENTS_CONTAINER]: locations.map((location) => location.id),
-  };
-  const dnd = useSidebarDnd(containers, (containerId, orderedIds) => {
-    if (containerId === AGENTS_CONTAINER) sidebarStore.setLocationOrder(orderedIds);
-    else sidebarStore.setGroupOrder(containerId, orderedIds);
-  });
-
-  function renderSessionGroup(locationId: string, roomKey: string, roomSessions: SessionStore[]) {
-    const container = agentSessionsContainer(locationId, roomKey);
-    const ordered = sidebarStore.orderGroupItems(
-      container,
-      roomSessions,
-      (session) => session.data.id,
-      true
-    );
-    containers[container] = ordered.map((session) => session.data.id);
-    const depth = 2;
-    return (
-      <SortableContext
-        items={ordered.map((session) => makeDndId(container, session.data.id))}
-        strategy={verticalListSortingStrategy}
-      >
-        {ordered.map((session) => (
-          <SortableLeaf key={session.data.id} id={makeDndId(container, session.data.id)}>
-            <SidebarSessionItem locationId={locationId} sessionId={session.data.id} depth={depth} />
-          </SortableLeaf>
-        ))}
-      </SortableContext>
-    );
-  }
-
+/** Render an agent's sessions grouped by room, under the agent row. */
+const AgentSessions = observer(function AgentSessions({
+  agentId,
+  locationId,
+  sessions,
+  depth,
+}: {
+  agentId: string;
+  locationId: string;
+  sessions: SessionStore[];
+  depth: number;
+}) {
   return (
-    <DndContext {...dnd}>
-      <SortableContext
-        items={locations.map((location) => makeDndId(AGENTS_CONTAINER, location.id))}
-        strategy={verticalListSortingStrategy}
-      >
-        {locations.map((location) => {
-          const locationId = location.id;
-          const expanded = sidebarStore.expandedLocationIds.has(locationId);
-          const allSessions = expanded ? sidebarStore.visibleSessionsForLocation(locationId) : [];
-          // Subagent sessions nest under their subagent row; the rest group by room.
-          const subagentSessions = allSessions.filter(isSubagentSession);
-          const grouped = groupByRoom(allSessions.filter((s) => !isSubagentSession(s)));
-          return (
-            <SortableBranch
-              key={locationId}
-              id={makeDndId(AGENTS_CONTAINER, locationId)}
-              header={<SidebarLocationItem locationId={locationId} depth={0} />}
-            >
-              {/* The agent's own rooms/sessions come first, then its subagents. */}
-              {expanded &&
-                grouped.map(([roomKey, roomSessions]) => {
-                  const groupKey = agentRoomGroupKey(locationId, roomKey);
-                  const roomExpanded = sidebarStore.isGroupExpanded(groupKey);
-                  return (
-                    <div key={roomKey}>
-                      <RoomRow
-                        label={roomLabel(roomKey)}
-                        count={roomSessions.length}
-                        expanded={roomExpanded}
-                        depth={1}
-                        bridgeType={switchRoomsStore.roomBridgeTypeById(roomKey)}
-                        onToggle={() => sidebarStore.toggleGroupExpanded(groupKey)}
-                        onOpenGateway={() => openRoomInGateway(roomKey)}
-                        onOpenChannel={
-                          switchRoomsStore.roomChannelUrl(roomKey)
-                            ? () => openRoomInMessagingApp(roomKey)
-                            : null
-                        }
-                      />
-                      {roomExpanded && renderSessionGroup(locationId, roomKey, roomSessions)}
-                    </div>
-                  );
-                })}
-              {expanded && (
-                <SidebarSubagentList
+    <>
+      {groupByRoom(sessions).map(([roomKey, roomSessions]) => {
+        // Sessions with no room sit directly under the agent, no room header.
+        if (roomKey === UNASSIGNED_ROOM_KEY) {
+          return roomSessions.map((session) => (
+            <SidebarSessionItem
+              key={session.data.id}
+              locationId={locationId}
+              sessionId={session.data.id}
+              depth={depth}
+            />
+          ));
+        }
+        const groupKey = agentRoomGroupKey(agentId, roomKey);
+        const roomExpanded = sidebarStore.isGroupExpanded(groupKey);
+        return (
+          <div key={roomKey}>
+            <RoomRow
+              label={roomLabel(roomKey)}
+              count={roomSessions.length}
+              expanded={roomExpanded}
+              depth={depth}
+              bridgeType={switchRoomsStore.roomBridgeTypeById(roomKey)}
+              onToggle={() => sidebarStore.toggleGroupExpanded(groupKey)}
+              onOpenGateway={() => openRoomInGateway(roomKey)}
+              onOpenChannel={
+                switchRoomsStore.roomChannelUrl(roomKey)
+                  ? () => openRoomInMessagingApp(roomKey)
+                  : null
+              }
+            />
+            {roomExpanded &&
+              roomSessions.map((session) => (
+                <SidebarSessionItem
+                  key={session.data.id}
                   locationId={locationId}
-                  sessions={subagentSessions}
-                  depth={1}
+                  sessionId={session.data.id}
+                  depth={depth + 1}
                 />
-              )}
-            </SortableBranch>
-          );
-        })}
-      </SortableContext>
-    </DndContext>
+              ))}
+          </div>
+        );
+      })}
+    </>
+  );
+});
+
+const AgentFocusedTree = observer(function AgentFocusedTree() {
+  const agents = scopedAgents();
+  return (
+    <>
+      {agents.map((entry) => {
+        const expanded = sidebarStore.isGroupExpanded(agentExpandKey(entry.agent.id));
+        return (
+          <div key={entry.agent.id}>
+            <SidebarAgentItem agent={entry.agent} depth={0} />
+            {expanded && (
+              <AgentSessions
+                agentId={entry.agent.id}
+                locationId={entry.agent.locationId}
+                sessions={agentSessions(entry)}
+                depth={1}
+              />
+            )}
+          </div>
+        );
+      })}
+    </>
   );
 });
 
 const RoomFocusedTree = observer(function RoomFocusedTree() {
-  // Collect every visible session across mounted agents, tagged with its agent.
-  const tagged: { locationId: string; session: SessionStore }[] = [];
-  for (const location of sidebarStore.filteredLocations) {
-    for (const session of sidebarStore.visibleSessionsForLocation(location.id)) {
-      tagged.push({ locationId: location.id, session });
+  // Tag every visible session with the agent it belongs to, then group by room.
+  const bySession = new Map<string, AgentEntry>();
+  const allSessions: SessionStore[] = [];
+  for (const entry of scopedAgents()) {
+    for (const session of agentSessions(entry)) {
+      bySession.set(session.data.id, entry);
+      allSessions.push(session);
     }
   }
 
-  const grouped = groupByRoom(tagged.map((t) => t.session));
-  const byKey = new Map(grouped);
-  const orderedRoomKeys = sidebarStore.orderRoomKeys(grouped.map(([roomKey]) => roomKey));
-
-  const containers: Record<string, string[]> = { [ROOMS_CONTAINER]: orderedRoomKeys };
-  const dnd = useSidebarDnd(containers, (containerId, orderedIds) => {
-    if (containerId === ROOMS_CONTAINER) sidebarStore.setRoomOrder(orderedIds);
-    else sidebarStore.setGroupOrder(containerId, orderedIds);
-  });
-
   return (
-    <DndContext {...dnd}>
-      <SortableContext
-        items={orderedRoomKeys.map((roomKey) => makeDndId(ROOMS_CONTAINER, roomKey))}
-        strategy={verticalListSortingStrategy}
-      >
-        {orderedRoomKeys.map((roomKey) => {
-          const roomSessions = byKey.get(roomKey) ?? [];
-          // Rooms in this view always have sessions, so default to expanded.
-          const roomViewKey = roomViewGroupKey(roomKey);
-          const expanded = sidebarStore.isGroupExpanded(roomViewKey);
-          const sessionIds = new Set(roomSessions.map((s) => s.data.id));
-          const byLocation = sidebarStore.filteredLocations
-            .map((location) => ({
-              locationId: location.id,
-              sessions: tagged
-                .filter((t) => t.locationId === location.id && sessionIds.has(t.session.data.id))
-                .map((t) => t.session),
-            }))
-            .filter((entry) => entry.sessions.length > 0);
-
-          return (
-            <SortableBranch
-              key={roomKey}
-              id={makeDndId(ROOMS_CONTAINER, roomKey)}
-              header={
-                <RoomRow
-                  label={roomLabel(roomKey)}
-                  count={roomSessions.length}
-                  expanded={expanded}
-                  depth={0}
-                  bridgeType={switchRoomsStore.roomBridgeTypeById(roomKey)}
-                  onToggle={() => sidebarStore.toggleGroupExpanded(roomViewKey)}
-                  onOpenGateway={() => openRoomInGateway(roomKey)}
-                  onOpenChannel={
-                    switchRoomsStore.roomChannelUrl(roomKey)
-                      ? () => openRoomInMessagingApp(roomKey)
-                      : null
-                  }
-                />
+    <>
+      {groupByRoom(allSessions).map(([roomKey, roomSessions]) => {
+        const roomViewKey = roomViewGroupKey(roomKey);
+        const expanded = sidebarStore.isGroupExpanded(roomViewKey);
+        // Agents that have a session in this room, in first-seen order.
+        const seen = new Set<string>();
+        const agentsInRoom: AgentEntry[] = [];
+        for (const session of roomSessions) {
+          const entry = bySession.get(session.data.id);
+          if (entry && !seen.has(entry.agent.id)) {
+            seen.add(entry.agent.id);
+            agentsInRoom.push(entry);
+          }
+        }
+        return (
+          <div key={roomKey}>
+            <RoomRow
+              label={roomLabel(roomKey)}
+              count={roomSessions.length}
+              expanded={expanded}
+              depth={0}
+              bridgeType={switchRoomsStore.roomBridgeTypeById(roomKey)}
+              onToggle={() => sidebarStore.toggleGroupExpanded(roomViewKey)}
+              onOpenGateway={() => openRoomInGateway(roomKey)}
+              onOpenChannel={
+                switchRoomsStore.roomChannelUrl(roomKey)
+                  ? () => openRoomInMessagingApp(roomKey)
+                  : null
               }
-            >
-              {expanded &&
-                byLocation.map((entry) => {
-                  // Reuse the agent row so it matches the agent-focused view
-                  // (icon, hover-chevron, sizing). Its expand state is the global
-                  // location expand, shared across the rooms an agent appears in.
-                  const agentExpanded = sidebarStore.expandedLocationIds.has(entry.locationId);
-                  // The agent's own sessions in this room render under it; the
-                  // subagents that have a session here render as sibling rows at
-                  // the same depth as the agent (not nested inside it).
-                  const parentSessions = entry.sessions.filter((s) => !isSubagentSession(s));
-                  const subagentSessions = entry.sessions.filter(isSubagentSession);
-                  const container = roomSessionsContainer(roomKey, entry.locationId);
-                  const orderedParent = sidebarStore.orderGroupItems(
-                    container,
-                    parentSessions,
-                    (session) => session.data.id,
-                    true
-                  );
-                  containers[container] = orderedParent.map((session) => session.data.id);
-                  return (
-                    <Fragment key={entry.locationId}>
-                      <SidebarLocationItem locationId={entry.locationId} depth={1} />
-                      {agentExpanded && (
-                        <SortableContext
-                          items={orderedParent.map((session) =>
-                            makeDndId(container, session.data.id)
-                          )}
-                          strategy={verticalListSortingStrategy}
-                        >
-                          {orderedParent.map((session) => (
-                            <SortableLeaf
-                              key={session.data.id}
-                              id={makeDndId(container, session.data.id)}
-                            >
-                              <SidebarSessionItem
-                                locationId={entry.locationId}
-                                sessionId={session.data.id}
-                                depth={2}
-                              />
-                            </SortableLeaf>
-                          ))}
-                        </SortableContext>
-                      )}
-                      {subagentSessions.length > 0 && (
-                        <SidebarSubagentList
-                          locationId={entry.locationId}
-                          depth={1}
-                          sessions={subagentSessions}
-                          onlyWithSessions
-                          groupSessionsByRoom={false}
-                        />
-                      )}
-                    </Fragment>
-                  );
-                })}
-            </SortableBranch>
-          );
-        })}
-      </SortableContext>
-    </DndContext>
+            />
+            {expanded &&
+              agentsInRoom.map((entry) => {
+                const agentSessionsHere = roomSessions.filter(
+                  (session) => bySession.get(session.data.id)?.agent.id === entry.agent.id
+                );
+                return (
+                  <Fragment key={entry.agent.id}>
+                    <SidebarAgentItem agent={entry.agent} depth={1} />
+                    {agentSessionsHere.map((session) => (
+                      <SidebarSessionItem
+                        key={session.data.id}
+                        locationId={entry.agent.locationId}
+                        sessionId={session.data.id}
+                        depth={2}
+                      />
+                    ))}
+                  </Fragment>
+                );
+              })}
+          </div>
+        );
+      })}
+    </>
   );
 });
